infra: add generic mbuf priv data
Some metadata are relevant for all nodes, such as the iface.
Modify the gr_mbuf structure layout to prepend a generic
set of metadata, in our case, a pointer to the interface.

This pointer was present in most of the existing metadata,
either as an "input_iface" (just named "iface"), or "output_iface".
Rename it, and use it either as input or output iface accordingly.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>