api: add external notifications
The API works with a request/response model.
This doesn't allow for asynchronous messages from Grout to the clients.

Add a mechanism for the client to request for notifications, as well as
the ability for any module to send a message.
The message type must be declared with the "EVENT_TYPE" macro to ensure
uniqueness in message types.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>