infra: enable pipeline mode
In order to handle the exception path, control_ouput node is used
to transfer packets from the worker nodes to the control path.
The callbacks running on the control thread do not expose any counters,
which hides a bit the packet processing from the graph.

Prepare the re-integration of the exception path handling as a node by
activating the pipeline mode.

In order to declare a node as an "exception node", which will run on the
control path threads, the new flag "GR_NODE_FLAG_CONTROL_PLANE" must be
set.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>