infra: enable pipeline mode
In order to handle the exception path, control_ouput node is used
to transfer packets from the worker nodes to the control path.
The callbacks running on the control thread do not expose any counters,
which hides a bit the packet processing from the graph.

Prepare the re-integration of the exception path handling as a node by
activating the pipeline mode.

In order to declare a node as an "exception node", which will run on the
control path threads, the new flag "GR_NODE_FLAG_CONTROL_PLANE" must be
set.

The graph create declares now 3 graphs that need to be instanciated and
destroyed at the same time:
- base graph: used as a template
- dataplane graph: which will be polled by the worker thread
- control plane graph: all ctl plane graph will be walked on by a single
  thread running on lcore 0.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>